fase7/requirements.txt
# Dashboard e Interface
streamlit==1.32.0
plotly==5.19.0

# Data Science e Machine Learning
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2

# Banco de Dados
SQLAlchemy==2.0.27
oracledb==2.0.1

# APIs e Requisições
requests==2.31.0

# Ambiente e Configuração
python-dotenv==1.0.1

# Utilitários
Pillow==10.2.0

